+++ CarDealerAPI/Class1.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace CarDealerAPIService.App
+namespace CarDealerAPI
 {
     public class Class1
     {
